more tests 2

+++ jqm-all/jqm-client/jqm-api-client-core/src/main/java/com/enioka/jqm/api/JobInstance.java
@@ -46,7 +46,7 @@ public class JobInstance
     private String user;
     private String sessionID;
     private State state;
-    private Integer position;
+    private Long position;
     private Queue queue;
     private String queueName;
     private String keyword1, keyword2, keyword3, definitionKeyword1, definitionKeyword2, definitionKeyword3, module, email, application;
@@ -127,12 +127,12 @@ public class JobInstance
      * Position in the queue. 0 if running.<br>
      * <strong>This is the value retrieved during the latest {@link JqmClient#getJob(int)} call and may not be up to date!</strong>
      */
-    public Integer getPosition()
+    public Long getPosition()
     {
         return position;
     }
 
-    void setPosition(Integer position)
+    void setPosition(Long position)
     {
         this.position = position;
     }
